Finished base character system with shared data between clients + character based authentication system

+++ client/src/client/Authentication/PlayerAuthentication.ts
@@ -4,14 +4,23 @@ import getUserCharacterData from "../PlayerMethods/getUserCharacterData";
 
 class PlayerAuthentication {
 	public static LoginCamera: Camera;
+	public static LocalPlayer: PlayerMp
 
 
 	constructor() {
 		PlayerAuthentication.LoginCamera = new Camera('loginCam', new mp.Vector3(-79.9, -1079.5, 310.2), new mp.Vector3(-74.8, -819.2, 326.2));
+		PlayerAuthentication.LocalPlayer = mp.players.local;
 
+		mp.events.add("render", PlayerAuthentication.handleUnauthed);
 		mp.events.add("playerReady", PlayerAuthentication.handlePlayerReady);
 		mp.events.add("client:loginStart", PlayerAuthentication.handleLoginStart);
 		mp.events.add("client:loginEnd", PlayerAuthentication.endClientLogin);
+	}
+
+	public static handleUnauthed() {
+		if (!getUserCharacterData()) {
+			PlayerAuthentication.LocalPlayer.freezePosition(true);
+		}
 	}
 
 	public static handleLoginStart() {
@@ -28,24 +37,29 @@ class PlayerAuthentication {
 	public static freezeAndBlurClient() {
 		mp.game.ui.displayRadar(false);
 		mp.gui.cursor.show(true, true);
-		mp.players.local.position = new mp.Vector3(-811.6, 174.9, 76.8);
-		mp.players.local.freezePosition(true);
-		mp.players.local.setAlpha(0);
+		PlayerAuthentication.LocalPlayer.position = new mp.Vector3(-811.6, 174.9, 76.8);
+		PlayerAuthentication.LocalPlayer.freezePosition(true);
+		PlayerAuthentication.LocalPlayer.setAlpha(0);
 		mp.game.cam.renderScriptCams(true, false, 0, true, false);
 		mp.game.graphics.transitionToBlurred(100);
-		mp.players.local.freezePosition(true);
+		PlayerAuthentication.LocalPlayer.freezePosition(true);
 	}
 
 	public static endClientLogin() {
+		mp.console.logInfo("" + JSON.stringify(getUserCharacterData()));
 		if (!getUserCharacterData()) return;
 
 		mp.events.call("browser:pushRouter", "/");
 		mp.game.ui.displayRadar(true);
-		mp.players.local.setAlpha(255);
+		PlayerAuthentication.LocalPlayer.setAlpha(255);
 		mp.game.graphics.transitionFromBlurred(100);
-		mp.players.local.freezePosition(false);
-		PlayerAuthentication.LoginCamera.delete();
+		PlayerAuthentication.LocalPlayer.freezePosition(false);
 		toggleChat(true);
+
+
+		if (PlayerAuthentication.LoginCamera) {
+			PlayerAuthentication.LoginCamera.delete();
+		}
 	}
 }
